🗣 Going Records with Java 17

+++ news-cast-explorer-common/src/main/java/org/jesperancinha/newscast/model/explorer/Author.java
@@ -9,6 +9,7 @@ import lombok.Setter;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
+import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
@@ -44,7 +45,7 @@ public class Author {
             cascade = CascadeType.MERGE)
     private final List<Message> messages = new ArrayList<>();
 
-    private Long nMessages;
+    private Integer nMessages;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "page_id",
